Allow RGB(A) images
This reworks the code to now handle (W,H), (W,H,3), (N,W,H), and
(N,W,H,3) images.  Additionally, the alpha component of RGBA is also
recognized so (W,H,4) and (N,W,H,4) are also recognized.  This also adds
a menu option in the image viewing area to reshape the array image.  The
user can now select the axis that are width, height, depth, and RGB(A).
This allows one to view the RGB layers separately or as a composite
image.