Store the index of the leafs in the color rows
We now store the QModelIndex of the leaves in the left hand combo box in
the color rows.  This is so we can dynamically update the `dbt_leaf` in
the cube math window.  At the moment, it can only establish the index
for leaves that have been exposed in the tree viewer.  This is because
ViTables does not assign an index to a node until it has been exposed.